# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Era Task API Documentation",
        "description": "API documentation for authentication and admin management.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get paginated list of campaigns",
                "operationId": "ef352ea3c85942668890a7bf6a14d840",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status (draft|queued|sending|done|failed)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of campaigns",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Campaign"
                ],
                "summary": "Create a new campaign",
                "operationId": "6393c021fc9cc2256183b3e14df01fd5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campaign created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateCampaignResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{id}/queue": {
            "post": {
                "tags": [
                    "Campaign"
                ],
                "summary": "Queue a campaign to start sending emails",
                "operationId": "b6254f169012f3d6f38fecbcb1a32a7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign queued",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "queued"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Campaign not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/unsubscribe/{signed}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Unsubscribe from marketing emails",
                "operationId": "d8114047ecc810aef424287ea2d2b7b2",
                "parameters": [
                    {
                        "name": "signed",
                        "in": "path",
                        "description": "Signed unsubscribe token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully unsubscribed"
                    },
                    "400": {
                        "description": "Invalid or expired link"
                    }
                },
                "security": []
            }
        },
        "/api/v1/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get a single campaign",
                "operationId": "24e491db356eef9f33f52197b9dcb535",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{id}/status": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaign statistics",
                "operationId": "09946e35cfd14dd5e4cbcc72c5caffb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CampaignStatsResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/v1/campaigns/{id}/test-send": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Send campaign template to a single email for testing",
                "operationId": "ff9df5048378434ce53c87da654d0efa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TestSendCampaignRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Test email sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TestSendCampaignResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CampaignListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CampaignStatsResponse"
                        }
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 100
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 7
                    }
                },
                "type": "object"
            },
            "CreateCampaignRequest": {
                "required": [
                    "subject",
                    "template_key"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "New Year Promo"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Happy New Year!"
                    },
                    "template_key": {
                        "type": "string",
                        "example": "discount"
                    },
                    "from_email": {
                        "type": "string",
                        "example": "admin@example.com"
                    },
                    "segment_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "filter_json": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "CreateCampaignResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "status": {
                        "type": "string",
                        "example": "draft"
                    },
                    "total_recipients": {
                        "type": "integer",
                        "example": 2841
                    }
                },
                "type": "object"
            },
            "CampaignStatsResponse": {
                "properties": {
                    "total_recipients": {
                        "type": "integer",
                        "example": 1000
                    },
                    "sent_count": {
                        "type": "integer",
                        "example": 800
                    },
                    "error_count": {
                        "type": "integer",
                        "example": 5
                    },
                    "status": {
                        "type": "string",
                        "example": "sending"
                    }
                },
                "type": "object"
            },
            "CampaignResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Holiday Promo"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Special Discount!"
                    },
                    "template_key": {
                        "type": "string",
                        "example": "promotion"
                    },
                    "from_email": {
                        "type": "string",
                        "example": "marketing@example.com"
                    },
                    "segment_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "status": {
                        "type": "string",
                        "example": "draft"
                    },
                    "total_recipients": {
                        "type": "integer",
                        "example": 1000
                    },
                    "sent_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "error_count": {
                        "type": "integer",
                        "example": 0
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2025-10-20T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2025-10-20T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "TestSendCampaignRequest": {
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "test@example.com"
                    }
                },
                "type": "object"
            },
            "TestSendCampaignResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Test email sent successfully"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "http://127.0.0.1:8000/oauth/token",
                        "refreshUrl": "http://127.0.0.1:8000oauth/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}